requirements.txt
Adafruit-Blinka==8.32.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-connectionmanager==1.0.0
adafruit-circuitpython-requests==3.0.0
adafruit-circuitpython-rgb-display==3.12.4
adafruit-circuitpython-typing==1.10.2
Adafruit-PlatformDetect==3.60.0
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.8.0
args==0.1.0
clint==0.5.1
Pillow==9.5.0
psutil==5.9.8
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
smbus==1.1.post2
sysv_ipc==1.1.0
typing_extensions==4.10.0
